Change point value for tests to allow doubles, adjust student test case scores to also contain score instead of pass/fail, add shell project scripts and point values to database

+++ server/src/main/java/edu/purdue/cs/encourse/service/ProfessorService.java
@@ -31,8 +31,8 @@ public interface ProfessorService {
     String listAllCommitsByTime(@NonNull String projectID);
     String listStudentCommitsByTime(@NonNull String projectID, @NonNull String userName);
 
-    int uploadTestScript(@NonNull String projectID, @NonNull String testName, @NonNull String testContents, boolean isHidden, int points);
-    int addTestScript(@NonNull String projectID, @NonNull String testName, boolean isHidden, int points);
+    int uploadTestScript(@NonNull String projectID, @NonNull String testName, @NonNull String testContents, boolean isHidden, double points);
+    int addTestScript(@NonNull String projectID, @NonNull String testName, boolean isHidden, double points);
     int modifyTestScript(@NonNull String projectID, @NonNull String testName, @NonNull String field, @NonNull String value);
     int runTestall(@NonNull String projectID);
     int runTestallForStudent(@NonNull String projectID, @NonNull String userName);
